requirements.txt
# amqp==5.2.0
aniso8601==9.0.1
anyio==4.3.0
asgiref==3.8.1
bcrypt==3.2.2
billiard==4.2.0
black==21.12b0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.1
cron-descriptor==1.4.3
distlib==0.3.8
Django==4.2.13
django-admin-honeypot==1.1.0
django-admin-soft-dashboard==1.0.12
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-environ==0.11.2
django-extensions==3.2.3
django-jazzmin==3.0.0
django-secure==1.0.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==25.2.0
filelock==3.14.0
flake8==3.9.2
flower==2.0.1
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==22.0.0
humanize==4.9.0
identify==2.5.36
idna==3.7
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.7
mccabe==0.6.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.7.0
pycparser==2.22
pyflakes==2.3.1
PyJWT==2.8.0
pytest==7.4.4
pytest-cov==2.12.1
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==3.5.3
sentry-sdk==1.45.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
watchfiles==0.21.0
wcwidth==0.2.13
whitenoise==5.3.0
